rag/pyproject.toml
[project]
name = "alzheimer_doc_rag"
version = "0.1.0"
description = "Rag based on alzheimer docs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "faiss-cpu>=1.12.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "openai>=1.99.9",
    "pydantic>=2.11.7",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.48.1",
    "wikipedia>=1.4.0",
]
